agents/sig-agents-system/sig-agents/requirements.txt
psycopg2-binary>=2.9.0
boto3>=1.34.0
botocore>=1.34.0
